setup image view

+++ app/src/main/java/com/kakapo/weatherapps/model/Sys.kt
@@ -7,23 +7,23 @@ data class Sys(
         val type: Int,
         val message: Double,
         val country: String?,
-        val sunnrise: Int,
-        val sunset: Int
+        val sunnrise: Long,
+        val sunset: Long
 ): Parcelable {
     constructor(parcel: Parcel) : this(
             parcel.readInt(),
             parcel.readDouble(),
             parcel.readString(),
-            parcel.readInt(),
-            parcel.readInt()) {
+            parcel.readLong(),
+            parcel.readLong()) {
     }
 
     override fun writeToParcel(parcel: Parcel, flags: Int) {
         parcel.writeInt(type)
         parcel.writeDouble(message)
         parcel.writeString(country)
-        parcel.writeInt(sunnrise)
-        parcel.writeInt(sunset)
+        parcel.writeLong(sunnrise)
+        parcel.writeLong(sunset)
     }
 
     override fun describeContents(): Int {
